Python/Templates/Web/ProjectTemplates/Python/Web/PollsFlaskJade/requirements.txt
Flask<1
pyjade<5
azure-storage<1
pymongo<3
